requirements/requirements-test.txt
pytest>=3.6
pytest-cov==2.4.0
pytest-xdist==1.15.0
pytest-localserver==0.3.7
pytest-mock==1.6.2
pytest-asyncio==0.10.0
asynctest==0.13.0
pytest-aiohttp==0.3.0